[v1.0.7 - 8.0.1] - Release
Added compatibility with Bazooka addon

Added compatibility with ChocolateBar Broker Display addon

Added compatibility with LibJostle-3.0 library and addons that use this library

Added new option that allows set the offset and scale of PetBattleFrameBar (the bottom battle pet bar)

Changed and improved the function that draws the background of the MultiBarBottomRightButtons (1 to 6). Now it no longer causes strange behaviors and don't produce taints.

Fixed the ExhaustionLevelFillBar from ExpBar (StatusBar) to set its size correctly

Fixed a bug with MicroButtons when the PetBattleFrameBar is showed

Fixed minor bugs